# backend-familleunie/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tontine Famille Unie API",
        "description": "API Laravel pour la gestion de la tontine Famille Unie",
        "contact": {
            "email": "admin@familleunie.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/contributions/me": {
            "get": {
                "tags": [
                    "Cotisations"
                ],
                "summary": "Contributions de l'utilisateur connecté (Paginé + Filtres)",
                "operationId": "f6196745f961ff24a8914ec75b21f388",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "default",
                                "3months",
                                "6months",
                                "all"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Historique des cotisations paginé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/funds/me": {
            "get": {
                "tags": [
                    "Cotisations"
                ],
                "summary": "Participations aux fonds de l'utilisateur connecté",
                "operationId": "3e7e7377950ab4885b4b657fab51006f",
                "responses": {
                    "200": {
                        "description": "Historique des fonds"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/contributions/summary": {
            "get": {
                "tags": [
                    "Cotisations"
                ],
                "summary": "Résumé des cotisations par réunion",
                "operationId": "29190f047fa7c86b537bcd907202909f",
                "responses": {
                    "200": {
                        "description": "Résumé par réunion"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/contributions": {
            "post": {
                "tags": [
                    "Cotisations"
                ],
                "summary": "Enregistrer une cotisation (ADMIN/TRESORIER)",
                "operationId": "ca7fd6ea7d3b77317b7be79b4fd901f2",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "contribution_type_id",
                                    "meeting_id",
                                    "parts"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "contribution_type_id": {
                                        "type": "integer"
                                    },
                                    "meeting_id": {
                                        "type": "integer"
                                    },
                                    "parts": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cotisation enregistrée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/contributions/{id}": {
            "patch": {
                "tags": [
                    "Cotisations"
                ],
                "summary": "Modifier une cotisation déjà enregistrée (ADMIN/TRESORIER)",
                "operationId": "11a1607478fe573b80fcf137199c30c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "contribution_type_id": {
                                        "type": "integer"
                                    },
                                    "meeting_id": {
                                        "type": "integer"
                                    },
                                    "parts": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cotisation mise à jour"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/fund-entries": {
            "post": {
                "tags": [
                    "Cotisations"
                ],
                "summary": "Enregistrer un dépôt dans une caisse (ADMIN/TRESORIER/COMMISSAIRE)",
                "operationId": "043642f99642cc1492bf9ec40cf28fd0",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "fund_type_id",
                                    "amount"
                                ],
                                "properties": {
                                    "fund_type_id": {
                                        "type": "integer"
                                    },
                                    "amount": {
                                        "type": "integer"
                                    },
                                    "meeting_id": {
                                        "type": "integer"
                                    },
                                    "direction": {
                                        "type": "string",
                                        "enum": [
                                            "in",
                                            "out"
                                        ]
                                    },
                                    "note": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dépôt enregistré"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/fund-entries/{id}": {
            "patch": {
                "tags": [
                    "Cotisations"
                ],
                "summary": "Modifier un mouvement de caisse déjà enregistré (ADMIN/TRESORIER/COMMISSAIRE)",
                "operationId": "7f87dfb9a75f482d4c6d55da8e4844c2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fund_type_id": {
                                        "type": "integer"
                                    },
                                    "member_id": {
                                        "type": "integer"
                                    },
                                    "amount": {
                                        "type": "integer"
                                    },
                                    "direction": {
                                        "type": "string",
                                        "enum": [
                                            "in",
                                            "out"
                                        ]
                                    },
                                    "note": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mouvement mis à jour"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/fund-entries": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Liste de tous les mouvements de caisse (ADMIN/TRESORIER/COMMISSAIRE)",
                "operationId": "fa3ffe497faeb56348ff9b5a24edba38",
                "parameters": [
                    {
                        "name": "fund_type_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "in",
                                "out"
                            ]
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mouvements de caisse paginés"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/contributions/report/{meetingId}": {
            "get": {
                "tags": [
                    "Cotisations"
                ],
                "summary": "Rapport des cotisations manquantes par réunion (ADMIN/TRESORIER)",
                "operationId": "f84ed7e0bfc6a04fc923a3724333a5d7",
                "parameters": [
                    {
                        "name": "meetingId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rapport des cotisations manquantes"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Entrez votre token Sanctum: Bearer {token}",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}